project/requirements.txt
# Core dependencies
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0

# Video processing
moviepy>=1.0.3
ffmpeg-python>=0.2.0
Pillow>=10.0.0

# Machine learning for bandit
mabwiser>=2.7.0
scikit-learn>=1.3.0

# Text processing
pyttsx3>=2.90  # for optional TTS
python-dotenv>=1.0.0

# Web scraping
selenium>=4.15.0
webdriver-manager>=4.0.0

# Scheduling
schedule>=1.2.0

# Image processing
opencv-python>=4.8.0

# Video downloading and processing
yt-dlp>=2023.7.6
ffprobe>=0.5

# API resilience and monitoring
tenacity>=8.2.0
prometheus-client>=0.17.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
ruff>=0.1.0
black>=23.0.0
isort>=5.12.0
